Inline BulkIngestService trait
